app/requirements.txt
setuptools
discord.py[voice]
Pillow